moved utils tests to utils
